requirements.txt
allure-pytest==2.9.45
pytest==7.4.0
webdriver-manager==3.8.0
selenium==4.11.2